src/main/webapp/app/entities/caja/update/caja-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize } from 'rxjs/operators';

import { CajaFormService, CajaFormGroup } from './caja-form.service';
import { ICaja } from '../caja.model';
import { CajaService } from '../service/caja.service';

@Component({
  selector: 'jhi-caja-update',
  templateUrl: './caja-update.component.html',
})
export class CajaUpdateComponent implements OnInit {
  isSaving = false;
  caja: ICaja | null = null;

  editForm: CajaFormGroup = this.cajaFormService.createCajaFormGroup();

  constructor(protected cajaService: CajaService, protected cajaFormService: CajaFormService, protected activatedRoute: ActivatedRoute) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ caja }) => {
      this.caja = caja;
      if (caja) {
        this.updateForm(caja);
      }
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const caja = this.cajaFormService.getCaja(this.editForm);
    if (caja.id !== null) {
      this.subscribeToSaveResponse(this.cajaService.update(caja));
    } else {
      this.subscribeToSaveResponse(this.cajaService.create(caja));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<ICaja>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(caja: ICaja): void {
    this.caja = caja;
    this.cajaFormService.resetForm(this.editForm, caja);
  }
}
